refactor: move cell into own component
